Got AlephTav working much better
