+++ MyHebrewBible/MyHebrewBible.Client/Components/PageHeader.razor
@@ -28,11 +28,27 @@
 
 	<div class="collapse navbar-collapse justify-content-end">
 
+		@* ToDo: a bit of a hack *@
+		@if (PageEnum == Nav.AlephTavs)
+		{
+			<div class="ps-1">
+				<MyHebrewBible.Client.Features.AlephTav.NavToAlephTavListButton />
+			</div>
+		}
+
+		@if (PageEnum == Nav.AlephTavList)
+		{
+			<div class="ps-1">
+				<MyHebrewBible.Client.Features.AlephTavList.NavBackToAlephTavsButton />
+			</div>
+		}
+
 		<div class="ps-1">
 			<a class="btn btn-primary" title="Configuration" href="">
 				<i class="fas fa-cog"></i>
 			</a>
 		</div>
+
 	</div>
 
 </nav>
