+++ MyHebrewBible/MyHebrewBible.Client/Components/VerseSections/WordSegmentsCard.razor
@@ -15,7 +15,7 @@ else
 				<button type="button" class="mt-1 btn btn-danger btn-sm" @onclick="ButtonCloseClick">
 					<i class="fas fa-times"></i>
 				</button>
-					<span class='float-end'>@BLB.nkjv.AnchorBCV(BookChapterVerse)</span> 
+					<span class='float-end'>@BLB.NKJV.AnchorBCV(BookChapterVerse)</span> 
 			</h6>
 		</div>
