+++ MyHebrewBible/MyHebrewBible.Client/Features/AlephTav/Enums/Filter.cs
@@ -10,16 +10,16 @@ public abstract class Filter : SmartEnum<Filter>
 	{
 		//internal const int All = 0;
 		internal const int Summary = 1;       
-		internal const int All613 = 2;
 		internal const int Anomalies = 3;
+		//internal const int All613 = 2;
 	}
 	#endregion
 
 	#region  Declared Public Instances
 	//public static readonly Filter All = new AllSE();
 	public static readonly Filter Summary = new SummarySE();
-	public static readonly Filter All613 = new All613SE();
 	public static readonly Filter Anomalies = new AnomaliesSE();
+	//public static readonly Filter All613 = new All613SE();
 	#endregion
 
 	private Filter(string name, int value) : base(name, value)  // Constructor
@@ -46,17 +46,19 @@ public abstract class Filter : SmartEnum<Filter>
 		public override string Title => "Summary";
 	}
 
-	private sealed class All613SE : Filter
-	{
-		public All613SE() : base($"{nameof(Id.All613)}", Id.All613) { }
-		public override string Title => "See all 634";
-	}
-
 	private sealed class AnomaliesSE : Filter
 	{
 		public AnomaliesSE() : base($"{nameof(Id.Anomalies)}", Id.Anomalies) { }
 		public override string Title => "Anomalies";
 	}
+
+	/*
+	private sealed class All613SE : Filter
+	{
+		public All613SE() : base($"{nameof(Id.All613)}", Id.All613) { }
+		public override string Title => "See all 634";
+	}
+	*/
 
 	#endregion
 }
