+++ MyHebrewBible/MyHebrewBible.Client/Features/AlephTav/NavToAlephTavListButton.razor
@@ -1,0 +1,9 @@
+﻿@using MyHebrewBible.Client.Enums
+
+<a class="btn btn-primary" title="@Nav.AlephTavList.Title" href="@Nav.AlephTavList.Index">
+	<i class="@Nav.AlephTavList.Icon"></i>
+</a>
+
+@code {
+
+}
