--- MyHebrewBible/MyHebrewBible.Client/Features/AlephTav/SummaryTable.razor
@@ -1,77 +1,0 @@
-﻿<h3><span class="text-danger">את</span> Standalone Aleph Tav List</h3>
-
-
-@* <div class="@GlobalEnums.MediaQuery.XsOrSm.DivClass">
-</div>
-
-<div class="@GlobalEnums.MediaQuery.MdOrLgOrXl.DivClass">
-</div>
- *@
-
-<div class="card border-success mb-2">
-	<div class="card-header">
-		Standalone Aleph Tav's <span class="text-danger">את</span> by book
-	</div>
-
-	<div class="card-body bg-light">
-		<p>
-			The table has two columns with a count of <i>Aleph Tav</i> occurrences for each book.
-			The first column is a count of the verses and the second column is the actual count of <i>Aleph Tav's</i>.
-			Since some verses have two <i>Aleph Tav's</i> in the same verse, that column will have a larger count.
-			The totals are listed on the bottom.
-		</p>
-	</div>
-
-</div>
-
-
-<div class="table-responsive">
-
-	<table class="table table-striped table-condensed">
-		<thead>
-			<tr>
-				<th><br /><br />Book</th>
-				<th>Verses<br />with<br />Aleph Tavs</th>
-				<th>Actual<br />Aleph Tav<br />Count</th>
-			</tr>
-		</thead>
-		<tbody>
-			@foreach (var item in Enums.AlephTavSummary.List.OrderBy(o => o.Value).ToList())
-			{
-				verseCnt += item.VerseCnt;
-				alephTavCnt += item.SatCount;
-				<tr>
-					<td>
-						@item.BibleBook.Title
-						@* <a asp-page="@Anchors.ParentIndex" asp-route-id="@item.Id" title="@item.Abrv">@item.Abrv</a> *@
-					</td>
-					<td>
-						@item.VerseCnt
-					</td>
-					<td>
-						@item.SatCount
-					</td>
-				</tr>
-			}
-		</tbody>
-		<tfoot class="table-group-divider">
-			<tr>
-				<td><b>Total</b></td>
-				<td><span class="text-danger"><b>@verseCnt</b></span></td>
-				<td><span class="text-danger"><b>@alephTavCnt</b></span></td>
-			</tr>
-		</tfoot>
-	</table>
-
-</div>
-
-
-@code {
-
-	// [Parameter, EditorRequired] public bool IsXs { get; set; }
-	// [Parameter, EditorRequired] public GlobalEnums.MediaQuery? MediaQuery { get; set; }
-
-	int verseCnt = 0;
-	int alephTavCnt = 0;
-
-}
