+++ MyHebrewBible/MyHebrewBible.Client/Features/AlephTavList/Index.razor
@@ -1,0 +1,35 @@
+﻿@page "/alephtavlist"
+
+@inject IToastService? Toast
+
+<PageHeader PageEnum="@MyHebrewBible.Client.Enums.Nav.AlephTavList" />
+
+<div class="@GlobalEnums.MediaQuery.XsOrSm.DivClass">
+	<BookChapterList OnBookAndChapterSelected="ReturnedBookAndChapter"
+									 MediaQuery="GlobalEnums.MediaQuery.XsOrSm" />
+</div>
+
+<div class="@GlobalEnums.MediaQuery.MdOrLgOrXl.DivClass">
+	<BookChapterList OnBookAndChapterSelected="ReturnedBookAndChapter"
+									 MediaQuery="GlobalEnums.MediaQuery.MdOrLgOrXl" />
+</div>
+
+@if (@CurrentBookAndChapter is not null)
+{
+	<LoadingProgress>
+		<WordPartByBookChapter BookAndChapter="CurrentBookAndChapter" />
+	</LoadingProgress>
+}
+
+@code {
+
+	public BookAndChapter? CurrentBookAndChapter { get; set; }
+
+	private void ReturnedBookAndChapter(BookAndChapter bookAndChapter)
+	{
+		//Toast!.ShowWarning($"Selected {bookAndChapter.ToString()}");
+		CurrentBookAndChapter = bookAndChapter!;
+	}
+
+}
+
