+++ MyHebrewBible/MyHebrewBible.Client/Features/AlephTavList/NavBackToAlephTavsButton.razor
@@ -1,0 +1,9 @@
+﻿@using MyHebrewBible.Client.Enums
+
+<a class="btn btn-outline-primary" title="@Nav.AlephTavs.Title" href="@Nav.AlephTavs.Index">
+	<i class="@Nav.AlephTavs.Icon"></i>
+</a>
+
+@code {
+
+}
