+++ MyHebrewBible/MyHebrewBible.Client/Features/AlephTavList/SummaryVM.cs
@@ -1,0 +1,5 @@
+﻿namespace MyHebrewBible.Client.Features.AlephTavList;
+
+public record SummaryVM(long ScriptureId, string BCV, long Chapter, long Verse, long HasTwo, int RowCnt);
+
+// Ignore Spelling: Bcv
