+++ MyHebrewBible/MyHebrewBible.Client/Features/AlephTavList/WordPartByBookChapter.razor
@@ -1,0 +1,46 @@
+﻿@using MyHebrewBible.Client.Helpers
+@using MyHebrewBible.Client.Enums
+
+<h3 class="text-center">Aleph Tav Verses</h3>
+
+@if (DetailWordList == null)
+{
+	<p><em>DetailWordList Loading...</em></p>
+}
+else
+{
+	if (SummaryVerseList == null)
+	{
+		<p><em>SummaryList is Loading...</em></p>
+	}
+	else
+	{
+		@foreach (var item in SummaryVerseList)
+		{
+			<div class="card  border-primary my-2">
+				<div class="card-body">
+					<h5 class="card-title ">
+						<span>@BLB.WLC.AnchorBCV(BookAndChapter!.BibleBook!.Value, (int)item.Chapter, (int)item.Verse)</span>
+						@if (item.HasTwo == 1) @* 1==true *@
+						{
+							<text>&nbsp;</text>
+							<span class="text-danger">את</span> <span class="text-danger">את</span>
+						}
+					</h5>
+
+					<p class="hebrew">
+						@foreach (var item2 in DetailWordList.Where(w => w.BCV == item.BCV))
+						{
+							@((MarkupString)HebrewFormat.GetSpan(item2.WordEnum, item2.Hebrew1, item2.Hebrew2, item2.Hebrew3))
+							@HebrewFormat.CheckMaqqef(item2.WordEnum, item2.Hebrew1, item2.Hebrew2, item2.Hebrew3)
+						}
+					</p>
+				</div>
+			</div>
+		}
+	}
+}
+
+
+
+
