+++ MyHebrewBible/MyHebrewBible.Client/Features/BookChapter/Verses/WordSegmentsCard.razor
@@ -13,7 +13,7 @@ else
 				<button type="button" class="mt-1 btn btn-danger btn-sm" @onclick="ButtonCloseClick">
 					<i class="fas fa-times"></i>
 				</button>
-				<span class='float-end'>@Enums.BLB.nkjv.AnchorBCV(BookAndChapter, (int)VerseNumber)</span>
+				<span class='float-end'>@Enums.BLB.NKJV.AnchorBCV(BookAndChapter, (int)VerseNumber)</span>
 			</h6>
 		</div>
