+++ MyHebrewBible/MyHebrewBible.Client/Features/Lists/Components/SubNavBar.razor
@@ -24,13 +24,14 @@
 				{
 					<li class="nav-item">
 						<a class="nav-link @GetActive(item)" href="@item.Page">
+							<i class="@item.Icon"></i>
 							<small>@item.Title</small>
 						</a>
 					</li>
 				}
 			</ul>
 
-			@*<i class="@item.Icon"></i>*@
+			
 
 
 		</div>
