+++ MyHebrewBible/MyHebrewBible.Client/Layout/HomeButton.razor
@@ -1,7 +1,5 @@
 ﻿@using Page = MyHebrewBible.Client.Enums.Nav
 
-@* btn-secondary  btn-sm *@
-
 <div class="d-sm-none">
 	<a class="btn btn-primary" title="<576px Xs" href="@Page.Home.Index">
 		<i class="@Page.Home.Icon"></i>
